requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.0
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
Click==7.0
colorama==0.4.3
crypto==1.4.1
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
Deprecated==1.2.10
distlib==0.3.0
Django==3.0.8
django-cors-headers==3.7.0
django-hosts==4.0
djangorestframework==3.12.2
Eel==0.14.0
filelock==3.0.12
Flask==1.1.1
Flask-Cors==3.0.8
Flask-MySQLdb==0.2.0
Flask-SocketIO==4.3.0
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
html5lib==1.0.1
humanfriendly==8.2
idna==2.9
importlib-metadata==1.6.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mechanize==0.4.5
mysqlclient==1.4.6
Naked==0.1.31
numpy==1.18.2
parso==0.8.1
pathlib==1.0.1
peewee==3.13.3
Pillow==7.2.0
pycparser==2.20
PyGithub==1.51
PyJWT==1.7.1
pylint==2.6.0
PyMySQL==0.9.3
pyparsing==2.4.7
pyquery==1.4.1
pyreadline==2.1
python-dateutil==2.8.1
python-engineio==3.13.0
python-magic==0.4.18
python-magic-bin==0.4.14
python-socketio==4.6.0
pytz==2020.1
pywin32==227
PyYAML==5.3.1
requests==2.23.0
response==0.4.0
robobrowser==0.5.3
scipy==1.4.1
selenium==3.141.0
shellescape==3.8.1
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tkintertable==1.3.2
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.8
validators==0.18.1
virtualenv==20.0.25
webencodings==0.5.1
Werkzeug==1.0.0
whichcraft==0.6.1
wrapt==1.12.1
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.2.0
